+++ include/H2DE/engine/H2DE_camera.h
@@ -251,7 +251,7 @@ public:
      * @param pauseSensitive If true, animation pauses when the game is paused.
      * @return Timeline controlling this animation.
      */
-    H2DE_Timeline* setTranslate(const H2DE_Translate& translate, uint32_t duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive = true);
+    H2DE_Timeline* setTranslate(const H2DE_Translate& translate, uint32_t duration, H2DE_Easing easing, const std::function<void()>& completed = nullptr, bool pauseSensitive = true);
     /**
      * @brief Animate the camera's game width to a new value over time.
      * 
@@ -264,7 +264,7 @@ public:
      * @param pauseSensitive If true, animation pauses when the game is paused.
      * @return Timeline controlling this animation.
      */
-    H2DE_Timeline* setGameWidth(float width, uint32_t duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive = true);
+    H2DE_Timeline* setGameWidth(float width, uint32_t duration, H2DE_Easing easing, const std::function<void()>& completed = nullptr, bool pauseSensitive = true);
     /**
      * @brief Animate the interface width over time.
      * 
@@ -277,7 +277,7 @@ public:
      * @param pauseSensitive If true, animation pauses when the game is paused.
      * @return Timeline controlling this animation.
      */
-    H2DE_Timeline* setInterfaceWidth(float width, uint32_t duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive = true);
+    H2DE_Timeline* setInterfaceWidth(float width, uint32_t duration, H2DE_Easing easing, const std::function<void()>& completed = nullptr, bool pauseSensitive = true);
 
     friend class H2DE_Engine;
     friend class H2DE_Renderer;